add better visualization of code

+++ 0002_generator_of_numbers.py
@@ -1,3 +1,3 @@
 def z3():
-    return [Decimal(20 + x * 5) / 10 for x in range(8)]
+    return [(20 + x * 5) / 10 for x in range(8)]
 print(z3())
